+++ .env
@@ -1,1 +1,1 @@
-VITE_BASE_URL = http://192.168.1.60:8080
+VITE_BASE_URL=https://turkey-prod.xquare.app
