+++ src/apis/index.ts
@@ -1,44 +1,13 @@
 import axios from "axios";
+import { CompletionResponse, ProblemsResponse } from "./type";
 import {
   CategoryDateDataRequest,
   CategoryDateDataResponse,
-  CompletionResponse,
   MarkingProblemRequest,
   MarkingProblemResponse,
-} from "./types";
+} from "./type";
 
 export const baseUrl = import.meta.env.VITE_BASE_URL;
-
-// 카테고리와 날짜별 조회 api
-export const CategoryDateData = async (category: string, date: string) => {
-  const requestData: CategoryDateDataRequest = { category, date };
-
-  try {
-    const response = await axios.post<CategoryDateDataResponse[]>(
-      `${baseUrl}/problem/category`,
-      requestData
-    );
-
-    console.log("성공 응답:", response.data);
-    return response.data;
-  } catch (err) {
-    console.error("전체 에러:", err);
-
-    if (axios.isAxiosError(err)) {
-      console.error("상태 코드:", err.response?.status);
-      console.error("응답 데이터:", err.response?.data);
-      console.error("요청 URL:", err.config?.url);
-      console.error("요청 파라미터:", err.config?.params);
-
-      throw new Error(
-        `API 요청 실패: ${err.response?.status || "NETWORK_ERROR"}`
-      );
-    }
-
-    console.error("알 수 없는 에러:", err);
-    throw new Error("알 수 없는 에러가 발생하였습니다.");
-  }
-};
 
 // 채점 api
 export const MarkingProblemData = async (id: string, userAnswer: string) => {
@@ -63,8 +32,6 @@ export const MarkingProblemData = async (id: string, userAnswer: string) => {
     throw err;
   }
 };
-
-import { CompletionResponse } from "./type";
 
 //완료 여부 전체 조회 api
 export const Complete = async () => {
@@ -109,4 +76,13 @@ export const CategoryDateData = async (category: string, date: string) => {
 
     console.error("알 수 없는 에러:", err);
   }
+};
+
+export const CheckAnswers = async (date: string) => {
+  const encodedDate = encodeURIComponent(date);
+  return await axios.get<ProblemsResponse>(`${baseUrl}/problem/${encodedDate}`);
+};
+
+export const CheckWrongAnswers = async () => {
+  return await axios.get<ProblemsResponse>(`${baseUrl}/problem/mypage`);
 };
