+++ src/apis/type.ts
@@ -1,10 +1,6 @@
 export type CompletionResponse = {
-  completions: Completions[];
-};
-
-export type Completions = {
   date: string;
-  isComplete: boolean;
+  complete: boolean;
 };
 
 export type ProblemsResponse = Problem[];
@@ -15,4 +11,30 @@ export interface Problem {
   isCorrect: boolean;
   answer: string;
   photoUrl: string;
+  problem: string;
+  problemDetail: string;
+  level: number;
 }
+
+export type CategoryDateDataRequest = {
+  category: string;
+  date: string;
+};
+
+export type CategoryDateDataResponse = {
+  id: number;
+  level: number;
+  problem: string;
+  problemDetail: string;
+  date: string;
+  photoUrl: string;
+};
+
+export type MarkingProblemRequest = {
+  id: string;
+  userAnswer: string;
+};
+
+export type MarkingProblemResponse = {
+  result: boolean;
+};
