+++ src/components/mathQuiz/MathQuiz2.tsx
@@ -4,7 +4,7 @@ import { Speaker, QuizCard } from "../../components";
 import NextButton from "../NextButton";
 import { InputAnswer } from "../../components/InputAnswer";
 import { backIcon } from "../../assets";
-import { CategoryDateDataResponse } from "../../apis/types";
+import { CategoryDateDataResponse } from "../../apis/type";
 import { MarkingProblemData } from "../../apis";
 import { useNavigate } from "react-router-dom";
